refactor: refactoring package name to fit MVC architecture
Refactoring contents
- README.md's feature list update
- rename package to fit MVC architecture (controller, model, view)
- create Display class that have print feature
