+++ src/main/java/baseball/controller/game/GameStatus.java
@@ -1,4 +1,4 @@
-package baseball.game;
+package baseball.controller.game;
 
 public enum GameStatus {
     NEW_GAME,
