+++ src/test/java/baseball/view/ScoreRenderTest.java
@@ -1,10 +1,7 @@
-package baseball.game;
+package baseball.view;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import baseball.game.score.ScoreBoard;
-import baseball.game.score.ScoreStatus;
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
@@ -13,16 +10,13 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class ScoreBoardTest {
-
-    private ScoreBoard scoreBoard = new ScoreBoard();
-
+public class ScoreRenderTest {
 
     @ParameterizedTest(name = "{index}: {2}")
     @MethodSource("validateParameters")
     @DisplayName("생성된 결과 문자열에 대한 유효성 검사")
-    public void validate_buildFromScore(int ball, int strike, String expected) throws Exception {
-        assertEquals(expected, buildResult(ballAndStrike(ball, strike)));
+    public void validate_buildFromScore(int ball, int strike, String expected) {
+        assertEquals(expected, ScoreRender.render(ballAndStrike(ball, strike)));
     }
 
     static Stream<Arguments> validateParameters() {
@@ -41,12 +35,5 @@ public class ScoreBoardTest {
         scoreStatusMap.put(ScoreStatus.STRIKE, strike);
 
         return scoreStatusMap;
-    }
-
-    private String buildResult(Map<ScoreStatus, Integer> scoreStatusMap) throws Exception {
-        Method buildScoreFromMethod = scoreBoard.getClass().getDeclaredMethod("buildScoreFrom", Map.class);
-        buildScoreFromMethod.setAccessible(true);
-
-        return (String)buildScoreFromMethod.invoke(scoreBoard, scoreStatusMap);
     }
 }
